+++ objc_encode/src/encoding.rs
@@ -28,9 +28,9 @@ pub enum Encoding<'a> {
     Class,
     Sel,
     Unknown,
-    BitField(u32),
+    BitField(u8),
     Pointer(&'a Encoding<'a>),
-    Array(u32, &'a Encoding<'a>),
+    Array(usize, &'a Encoding<'a>),
     Struct(&'a str, &'a [Encoding<'a>]),
     Union(&'a str, &'a [Encoding<'a>]),
 }
